python-server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain-openai
langchain-core
openai
